static cast size()

+++ test/index/vocabulary/PolymorphicVocabularyTest.cpp
@@ -38,14 +38,15 @@ void testForVocabType(VocabularyType::Enum vocabType) {
   wI = vocab.upper_bound("gamma", ql::ranges::less{});
   EXPECT_TRUE(wI.isEnd());
 
-  EXPECT_EQ(std::visit([](auto& u) { return u.size(); },
+  EXPECT_EQ(std::visit([](auto& u) { return static_cast<uint64_t>(u.size()); },
                        vocab.getUnderlyingVocabulary()),
             3);
 
   const auto& vocabConst = vocab;
-  EXPECT_EQ(std::visit([](const auto& u) { return u.size(); },
-                       vocabConst.getUnderlyingVocabulary()),
-            3);
+  EXPECT_EQ(
+      std::visit([](const auto& u) { return static_cast<uint64_t>(u.size()); },
+                 vocabConst.getUnderlyingVocabulary()),
+      3);
 }
 }  // namespace
